2.1

+++ BALANCE_ DOWN/APP/RTOS/Src/control_task.c
@@ -18,11 +18,12 @@ void control_task(void)
 	{
 		balance_chassis_task();
 		can_bus_send_task();
+		
 	}
 	
 	if(time_tick%10==9)
 	{
-		power_data_set_handle(CAN2,b_chassis.Max_power_to_PM01);
+		power_data_set_handle(CAN2,(uint16_t)(b_chassis.Max_power_to_PM01));
 	}
 	if(time_tick%10==5)
 	{
